Fix end-of-program line not printed error
OR-list and AND-list is NOT a same command and can not be mixed like one, currently:

```bash
	if meta_util_is_parameter_set_and_not_null META_APPLICATION_NAME\
		|| meta_util_is_parameter_set_and_not_null META_APPLICATION_DEVELOPER_NAME\
		|| meta_util_is_parameter_set_and_not_null META_APPLICATION_SITE_URL\
		|| meta_util_is_parameter_set_and_not_null META_PROGRAM_LICENSE\
		|| meta_util_is_parameter_set_and_not_null META_PROGRAM_PAUSE_BEFORE_EXIT\
			&& [ "${META_PROGRAM_PAUSE_BEFORE_EXIT}" -eq 1 ]; then
		printf -- "------------------------------------\n"
	fi
```

First, lists are left-associative and have no precedence difference between AND-list and OR-list.  Don't mistaken it with the logical AND OR operator in BASHDOC:Basic Shell Features » Shell Commands » Compound Commands » Conditional Constructs

The previous OR-list turned ENTIRELY into TRUE(0) after any of it member resolves to TRUE(0), destroying the `meta_util_is_parameter_set_and_not_null META_PROGRAM_PAUSE_BEFORE_EXIT` and directory compares to the next AND-list member [ "${META_PROGRAM_PAUSE_BEFORE_EXIT}" -eq 1 ], so the line will be printed ONLY when META_PROGRAM_PAUSE_BEFORE_EXIT is not null and expands to "1".

This commit fixes the issue by using command grouping (BASHDOC:Basic Shell Features » Shell Commands » Compound Commands » Grouping Commands), which groups the:

```
meta_util_is_parameter_set_and_not_null META_PROGRAM_PAUSE_BEFORE_EXIT\
			&& [ "${META_PROGRAM_PAUSE_BEFORE_EXIT}" -eq 1 ]
```

and list part so it can become a member of the OR-list, fixes the logic

Signed-off-by: 林博仁 <Buo.Ren.Lin@gmail.com>